refactor: external calls
