+++ packages/sdk/src/External.ts
@@ -1,0 +1,10 @@
+export * as AaveV2 from "@enzymefinance/sdk/internal/External/AaveV2";
+export * as BalancerV2 from "@enzymefinance/sdk/internal/External/BalancerV2";
+export * as CompoundV2 from "@enzymefinance/sdk/internal/External/CompoundV2";
+export * as Convex from "@enzymefinance/sdk/internal/External/Convex";
+export * as Curve from "@enzymefinance/sdk/internal/External/Curve";
+export * as IdleV4 from "@enzymefinance/sdk/internal/External/IdleV4";
+export * as Liquity from "@enzymefinance/sdk/internal/External/AaveV2";
+export * as Maple from "@enzymefinance/sdk/internal/External/Maple";
+export * as TheGraph from "@enzymefinance/sdk/internal/External/TheGraph";
+export * as UniswapV2 from "@enzymefinance/sdk/internal/External/UniswapV2";
