+++ packages/sdk/src/internal/Extensions/ExternalPositions/Liquity.ts
@@ -1,6 +1,5 @@
-import { Viem } from "@enzymefinance/sdk/Utils";
 import * as ExternalPositionManager from "@enzymefinance/sdk/internal/ExternalPositionManager";
-import { type Address, type Hex, type PublicClient, decodeAbiParameters, encodeAbiParameters } from "viem";
+import { type Address, type Hex, decodeAbiParameters, encodeAbiParameters } from "viem";
 
 export type Action = typeof Action[keyof typeof Action];
 export const Action = {
@@ -250,105 +249,4 @@ export function repayBorrowDecode(encoded: Hex): RepayBorrowArgs {
     upperHint,
     lowerHint,
   };
-}
-
-//--------------------------------------------------------------------------------------------
-// EXTERNAL CONTRACT METHODS
-//--------------------------------------------------------------------------------------------
-
-const troveManagerAbi = {
-  inputs: [
-    {
-      internalType: "address",
-      name: "",
-      type: "address",
-    },
-  ],
-  name: "Troves",
-  outputs: [
-    {
-      internalType: "uint256",
-      name: "debt",
-      type: "uint256",
-    },
-    {
-      internalType: "uint256",
-      name: "coll",
-      type: "uint256",
-    },
-    {
-      internalType: "uint256",
-      name: "stake",
-      type: "uint256",
-    },
-    {
-      internalType: "enum TroveManager.Status",
-      name: "status",
-      type: "uint8",
-    },
-    {
-      internalType: "uint128",
-      name: "arrayIndex",
-      type: "uint128",
-    },
-  ],
-  stateMutability: "view",
-  type: "function",
-} as const;
-
-type Trove = {
-  debt: bigint;
-  collateral: bigint;
-  stake: bigint;
-  status: number;
-  arrayIndex: bigint;
-};
-
-export async function getTrove(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    liquityTroveManager: Address;
-    debtPosition: Address;
-  }>,
-) {
-  const [debt, collateral, stake, status, arrayIndex] = await Viem.readContract(client, args, {
-    abi: [troveManagerAbi],
-    functionName: "Troves",
-    address: args.liquityTroveManager,
-    args: [args.debtPosition],
-  });
-
-  return {
-    debt,
-    collateral,
-    stake,
-    status,
-    arrayIndex,
-  };
-}
-
-export async function getTroves(
-  client: PublicClient,
-  args: Viem.ContractCallParameters<{
-    liquityTroveManager: Address;
-    debtPositions: [Address];
-  }>,
-) {
-  const troves = await Promise.all(
-    args.debtPositions.map(async (position) => {
-      const trove = await getTrove(client, {
-        ...args,
-        debtPosition: position,
-      });
-
-      return { position, trove };
-    }),
-  );
-
-  const troveMap: Record<Address, Trove> = {};
-  for (const { position, trove } of troves) {
-    troveMap[position] = trove;
-  }
-
-  return troveMap;
 }
